+++ src/mds/models.py
@@ -1,6 +1,6 @@
 from gino.ext.starlette import Gino
 from sqlalchemy.dialects.postgresql import JSONB
-from sqlalchemy import Integer, Sequence
+from sqlalchemy import BigInteger, Sequence
 
 from . import config
 
@@ -28,4 +28,4 @@ class MetadataIndex(db.Model):
 
     id_sec = Sequence(__tablename__ + "_id_seq")
     guid = db.Column(db.Unicode(), primary_key=True)
-    id = db.Column(Integer, id_sec, server_default=id_sec.next_value())
+    id = db.Column(BigInteger, id_sec, server_default=id_sec.next_value())
